refactor ssh key generation
